chore: more form array work, ensure new array values are created properly
