+++ app/ui-react/packages/auto-form/src/FormBuilder.tsx
@@ -4,10 +4,10 @@ import {
   IFormArrayControlProps,
   IFormControlProps,
   IFormDefinition,
-  IFormDefinitionProperty,
   INamedConfigurationProperty,
   IRenderFieldProps,
 } from './models';
+import { enrichAndOrderProperties, massageType, sanitizeValues } from './utils';
 import {
   FormCheckboxComponent,
   FormDurationComponent,
@@ -39,7 +39,7 @@ export class FormBuilder<T> extends React.Component<IFormBuilderProps<T>> {
    */
   public getField = (props: IRenderFieldProps): any => {
     // Massage the value in the definition to an input type
-    const type = this.massageType(props.property);
+    const type = massageType(props.property);
     const componentTypemaps = {
       checkbox: FormCheckboxComponent,
       duration: FormDurationComponent,
@@ -86,157 +86,15 @@ export class FormBuilder<T> extends React.Component<IFormBuilderProps<T>> {
   };
 
   public render() {
-    const propertiesArray = this.enrichAndOrderProperties(
-      this.props.definition
-    );
-    const massagedValue = this.sanitizeValues(
+    const propertiesArray = enrichAndOrderProperties(this.props.definition);
+    const massagedValue = sanitizeValues(
       this.props.definition,
       this.props.initialValue
     );
     return this.props.children({
       getField: this.getField,
-      initialValue: massagedValue,
+      initialValue: massagedValue as T,
       propertiesArray,
     });
-  }
-
-  /**
-   * Ensure that the input values match the property definitions
-   */
-  private sanitizeValues(definition: IFormDefinition, initialValue: any): T {
-    return Object.keys(definition).reduce((result, key): any => {
-      const prop = definition[key];
-      let value = this.massageValue(prop, initialValue[key]);
-      if (value == null) {
-        value = this.massageValue(prop, prop.defaultValue);
-      }
-      return { ...result, [key]: value };
-    }, {}) as T;
-  }
-
-  /**
-   * Add the 'name' field from the property ID and sort them by the 'order' property
-   */
-  private enrichAndOrderProperties(definition: IFormDefinition) {
-    return Object.keys(definition)
-      .map(key => ({
-        ...definition[key],
-        name: key,
-        required: this.massageRequired(definition[key]),
-        type: this.massageType(definition[key]),
-      }))
-      .sort((a, b) => {
-        const aOrder = (a.order || 0) as number;
-        const bOrder = (b.order || 0) as number;
-        return aOrder - bOrder;
-      });
-  }
-
-  /**
-   * Converts various values passed into the property type to known input types
-   *
-   * @param property
-   */
-  private massageType(property: IFormDefinitionProperty) {
-    let type = property.type || 'text';
-    switch (type) {
-      case 'int':
-      case 'integer':
-      case 'long':
-        type = 'number';
-        break;
-      case 'string':
-        type = 'text';
-        break;
-      case 'boolean':
-        type = 'checkbox';
-    }
-    if (typeof property.enum !== 'undefined' && property.enum.length) {
-      type = 'select';
-    }
-    if (typeof property.secret === 'boolean') {
-      type = 'password';
-    }
-    return type;
-  }
-
-  /**
-   * Ensure that the 'required' property is false for checkboxes and hidden fields
-   *
-   * This is a candidate for removal in the future, as it's a workaround
-   *
-   * @param property
-   */
-  private massageRequired(property: IFormDefinitionProperty): any {
-    switch (property.type) {
-      case 'boolean':
-      case 'checkbox':
-      case 'hidden':
-        return false;
-      default:
-        return property.required;
-    }
-  }
-
-  private getArrayRows(missing: number, definition: IFormDefinition) {
-    const answer: any[] = [];
-    for (let i = 0; i < missing; i++) {
-      answer.push({});
-    }
-    return answer;
-  }
-
-  private sanitizeInitialArrayValue(
-    definition: IFormDefinition,
-    value?: any[],
-    minimum?: number
-  ) {
-    const sanitizedValue = value || [];
-    const available = sanitizedValue.length;
-    const missing = (minimum || 0) - available;
-    if (missing < 0) {
-      return value;
-    }
-    return [...sanitizedValue, ...this.getArrayRows(missing, definition)];
-  }
-
-  /**
-   * Converts the given value from a string to the type defined in the property definition
-   *
-   * This is a candidate for removal as it's a workaround
-   *
-   * @param property
-   * @param value
-   */
-  private massageValue(property: IFormDefinitionProperty, value?: any) {
-    switch (property.type) {
-      case 'number':
-        return parseInt(value, 10);
-      case 'boolean':
-      case 'checkbox':
-        return String(value).toLocaleLowerCase() === 'true';
-      case 'array':
-        const minElements =
-          typeof property.arrayDefinitionOptions !== 'undefined'
-            ? property.arrayDefinitionOptions.minElements
-            : 0;
-        return this.sanitizeInitialArrayValue(
-          property.arrayDefinition || {},
-          value,
-          minElements
-        );
-      default:
-        // if the value has an enum property
-        // a select control is used, default
-        // to the first available value
-        if (
-          typeof value === 'undefined' &&
-          property.enum &&
-          property.enum.length > 0
-        ) {
-          return property.enum[0].value;
-        }
-        return value;
-    }
   }
 }
