+++ app/ui-react/syndesis/src/modules/settings/pages/OAuthAppsPage.tsx
@@ -17,7 +17,6 @@ import {
 } from '@syndesis/ui';
 import {
   allFieldsRequired,
-  applyInitialValues,
   getRequiredStatusText,
   toFormDefinition,
   validateConfiguredProperties,
@@ -196,10 +195,8 @@ export class OAuthAppsPage extends React.Component<{}, IOAuthAppsPageState> {
                                           const configured =
                                             typeof oauthApp.configuredProperties !==
                                             'undefined';
-                                          const configuration = applyInitialValues(
-                                            definition,
-                                            oauthApp.configuredProperties
-                                          );
+                                          const configuration =
+                                            oauthApp.configuredProperties;
                                           const key = JSON.stringify(
                                             configuration
                                           );
@@ -255,7 +252,9 @@ export class OAuthAppsPage extends React.Component<{}, IOAuthAppsPageState> {
                                                     values
                                                   )
                                                 }
-                                                initialValue={configuration}
+                                                initialValue={
+                                                  configuration as IFormValue
+                                                }
                                                 onSave={(
                                                   configuredProperties,
                                                   actions
